refactor: CookieUtils로 정리
