--- frontend/src/utils/hasCookie.ts
@@ -1,4 +1,0 @@
-export const hasCookie = (name: string) => {
-  const cookies = document.cookie.split(';').map((cookie) => cookie.trim());
-  return cookies.some((cookie) => cookie.startsWith(`${name}=`));
-};
